v1.2.0 – Add support for both blur and keyup events. (#31)
We've added f-validate to a new project, however there's a few things we need to add before we get feature parity with solution we're replacing.

The previous solution uses [jQuery Validate](https://github.com/jquery-validation/jquery-validation/). The default behaviour of this library is to delay validation until the input blur event has fired once. This provides an UX improvement of not seeing a validation error before the user has finished typing. This is most useful when the validation type is email or min-length. This PR attempts to recreate that logic.

A few things:

- I couldn't think of a better name than `hybridMode` which I appreciate isn't great 😄 
- I've tried to make minimal changes to the existing code
- There's not really any tests around any of this yet. If this approach looks OK, I'll add additional tests to cover these scenarios


## UI Review Checks

- [x] UI Documentation has been [updated]
- [x] JavaScript Tests have been [updated]
- [x] Module has been tested in Global Documentation

## Browsers Tested

- [x] Chrome
- [x] Edge
- [ ] Internet Explorer 11
- [x] Mobile (i.e. iPhone/Android - please list device)

### List any other browsers that this PR has been tested in:

- [View our supported browser list](https://fozzie.just-eat.com/documentation/general/browser-support)